방향성 검토 실시간 -> 배치로 S3에 HTML 업로드

+++ weekly/src/main/java/kr/damda/weekly/dto/response/SingleResponseDto.java
@@ -15,5 +15,5 @@ public class SingleResponseDto {
 
     private Integer validateDeviceCount;
 
-    private Long dataVolume;
+    private double dataVolume;
 }
